common: Handle case where httpd is not running.
For example if in setup_file() we skip all of the tests for a particular
version.  This was causing silent failures in CI.

+++ suite/common.bash
@@ -115,12 +115,14 @@ start_httpd()
 }
 stop_httpd()
 {
-	local pid=$(<${HTTPD_PID})
-	if [ -n "${pid}" ]; then
-		kill ${pid}
-		wait ${pid} 2>/dev/null || true
+	if [ -f ${HTTPD_PID} ]; then
+		local pid=$(<${HTTPD_PID})
+		if [ -n "${pid}" ]; then
+			kill ${pid}
+			wait ${pid} 2>/dev/null || true
+		fi
+		rm -f ${HTTPD_PID}
 	fi
-	rm -f ${HTTPD_PID}
 }
 
 #
